Day 1 Julia Typing correction

+++ julia/day01.jl
@@ -4,7 +4,7 @@ function get_inputs(fn::String="toy")::Array{Int64}
 end
 
 
-function get_two(fn::String="toy")
+function get_two(fn::String="toy")::Int64
     s = 1 
     for i in [x for x in get_inputs(fn) if (2020 - x) in get_inputs(fn)]
         s = s * i
@@ -15,10 +15,9 @@ end
 @assert(get_two("toy") == 514579)
 @assert(get_two("puzzle") == 889779)
 
-function get_three(fn::String="toy")
+function get_three(fn::String="toy")::Int64
     s = 1 
     k = get_inputs(fn)
-    # println(unique([x for x in k for y in k if (2020 - x - y) in k]))
 
     for i in unique([x for x in k for y in k if (2020 - x - y) in k])
         s = s * i
